possibility to run without drbd, e.g. no cluster, just mysql master-slave on different hosts possibility to go without a temporary volume, but copy directly from snapshot

+++ mysql-replication-fix-lvm-drbd-ha-vserver.sh
@@ -8,6 +8,7 @@
 #exit 1
 
 SEITEB="otherside"
+NOTMPLV="0"
 TMPLV="temp"
 TMPVG=""
 VGSEITEB=""
@@ -16,10 +17,13 @@ SIZE=""
 SIZESNAP=""
 MASTER=""
 SLAVE=""
+DRBD="1"
 REPLINFO="/root/master-replication-info"
 
 print_help() {
 	echo "Usage: $0 --seiteb=otherside --tmplv=temp --tmpvg=vg --vgseiteb=vgname --lv=lvname --size=size --sizesnap=sizesnap --master=master --slave=slave --replinfo=replinfo" >&2
+	echo "--nodrbd: no drbd involved" >&2
+	echo "--notmplv: no tmplv, copy directly from snapshot" >&2
 	echo "--seiteb: hostname/ip of other cluster host, defaults to 'otherside', make sure passwordless login works with ssh" >&2
 	echo "--tmplv:  name of temporary logical volume on $SEITEB, defaults to 'temp'" >&2
 	echo "--tmpvg:  name of volumegroup for temporary logical volume on $SEITEB, defaults to 'vg_$SEITEB'" >&2
@@ -37,12 +41,14 @@ error() {
   exit 1
 }
 
-TEMP=`getopt -o h --long help,seiteb:,tmplv:,tmpvg:,vgseiteb:,lv:,size:,sizesnap:,master:,slave:,fs:,replinfo: -- "$@"`
+TEMP=`getopt -o h --long help,nodrbd,seiteb:,tmplv:,tmpvg:,vgseiteb:,lv:,size:,sizesnap:,master:,slave:,fs:,replinfo: -- "$@"`
 if [ $? != 0 ] ; then print_help >&2 ; exit 1 ; fi
 eval set -- "$TEMP"
 while true ; do
 	case "$1" in
+		--nodrbd) DRBD=0; shift 2;;
 		--seiteb) SEITEB=$2; shift 2;;
+		--notmplv) NOTMPLV=0; shift 2;;
 		--tmplv) TMPLV=$2; shift 2;;
 		--tmpvg) TMPVG=$2; shift 2;;
 		--vgseiteb) VGSEITEB=$2; shift 2;;
@@ -79,8 +85,8 @@ else
 		error "$SEITEB nicht per ssh erreichbar, parameter --seiteb"
 	fi
 fi
-if [ x$TMPLV == x ]; then
-	error "Missing parameter --tmplv"
+if [ x$NOTMPLV == x0 ] && [ x$TMPLV == x ]; then
+	error "Missing parameter --tmplv and not using --notmplv"
 else
 	if [ x$TMPVG == x ]; then
 		TMPVG=vg_$SEITEB
@@ -130,8 +136,10 @@ fi
 
 echo "Parameters:
 SEITEB=\"$SEITEB\"
+NOTMPLV=\"$NOTMPLV\"
 TMPLV=\"$TMPLV\"
 TMPVG=\"$TMPVG\"
+DRBD=\"$DRBD\"
 VGSEITEB=\"$VGSEITEB\"
 LV=\"$LV\"
 SIZE=\"$SIZE\"
@@ -150,13 +158,25 @@ if [ "x$GO" != "xYes, I understand this might destroy all my data" ]; then
 fi
 
 # auf $SEITEA, 1 shell sparen. $SEITEB damit der Snapshot schnell wieder aufgeloest werden kann als Puffer
-ssh -X -a $SEITEB "(
-lvcreate -L$SIZE -n $TMPLV $TMPVG
-mkfs.ext4 /dev/$TMPVG/$TMPLV
-mkdir -p /mnt/tempmysqlsync
-mkdir -p /mnt/tempmysqlsnap
-mount /dev/$TMPVG/$TMPLV /mnt/tempmysqlsync
-)"
+if [ x$DRBD == x1 ]; then
+	ssh -X -a $SEITEB "(
+    if [ x$NOTMPLV == x0 ]; then
+		  lvcreate -L$SIZE -n $TMPLV $TMPVG
+		  mkfs.ext4 /dev/$TMPVG/$TMPLV
+		  mkdir -p /mnt/tempmysqlsync
+		  mount /dev/$TMPVG/$TMPLV /mnt/tempmysqlsync
+    fi
+		mkdir -p /mnt/tempmysqlsnap
+	)"
+else
+  if [ x$NOTMPLV == x0 ]; then
+		lvcreate -L$SIZE -n $TMPLV $TMPVG
+		mkfs.ext4 /dev/$TMPVG/$TMPLV
+		mkdir -p /mnt/tempmysqlsync
+		mount /dev/$TMPVG/$TMPLV /mnt/tempmysqlsync
+  fi
+	mkdir -p /mnt/tempmysqlsnap
+fi
 
 MOUNTPOINT=$(df /etc/vservers/$MASTER/vdir/var/lib/mysql | sed '1d; s/.* //')
 RELPATH=$(echo $(cd -P /etc/vservers/$MASTER/vdir/var/lib/mysql; pwd) | sed "s#^$MOUNTPOINT##")
@@ -175,38 +195,80 @@ echo sync2 done
 fsfreeze -f $MOUNTPOINT
 echo fsfreeze done
 sleep 10
-ssh -X -a $SEITEB lvcreate -s -L$SIZESNAP -n tempmysqlsnap /dev/$VGSEITEB/$LV
+if [ x$DRBD == x1 ]; then
+  ssh -X -a $SEITEB lvcreate -s -L$SIZESNAP -n tempmysqlsnap /dev/$VGSEITEB/$LV
+else
+  lvcreate -s -L$SIZESNAP -n tempmysqlsnap /dev/$VGSEITEA/$LV
+fi
 echo snap done
 fsfreeze -u $MOUNTPOINT
 echo fsunfreeze done
 vserver $MASTER exec bash -c "(echo > /tmp/replicationpipe; rm /tmp/replicationpipe)"
 
-ssh -X -a $SEITEB "(
-mount -t $FS /dev/$VGSEITEB/tempmysqlsnap /mnt/tempmysqlsnap &&
-( cd /mnt/tempmysqlsnap/$RELPATH/ && tar cf - . ) | mbuffer -m 8G | ( cd /mnt/tempmysqlsync && tar xvpf - ) &&
-umount /mnt/tempmysqlsnap &&
-lvremove -f /dev/$VGSEITEB/tempmysqlsnap
-)"
+if [ x$DRBD == x1 ]; then
+	ssh -X -a $SEITEB "(
+	  mount -t $FS /dev/$VGSEITEB/tempmysqlsnap /mnt/tempmysqlsnap &&
+	  ( if [ x$NOTMPLV == x0 ]; then
+	      cd /mnt/tempmysqlsnap/$RELPATH/ && tar cf - . ) |
+	      mbuffer -m 8G |
+	      ( cd /mnt/tempmysqlsync && tar xvpf - ) &&
+	      umount /mnt/tempmysqlsnap &&
+	      lvremove -f /dev/$VGSEITEB/tempmysqlsnap
+	    fi
+	  )"
+else
+  mount -t $FS /dev/$VGSEITEA/tempmysqlsnap /mnt/tempmysqlsnap &&
+  if [ x$NOTMPLV == x0 ]; then
+	  cd /mnt/tempmysqlsnap/$RELPATH/ && tar cf - . ) |
+	  mbuffer -m 8G |
+	  ( cd /mnt/tempmysqlsync && tar xvpf - ) &&
+    umount /mnt/tempmysqlsnap &&
+    lvremove -f /dev/$VGSEITEB/tempmysqlsnap
+  fi
+fi
 
 MASTERLOGFILE=$(sed '1d; s/\t.*//' /etc/vservers/$MASTER/vdir$REPLINFO)
 MASTERLOGPOS=$(sed '1d; s/^[^\t]*\t//; s/\t//g' /etc/vservers/$MASTER/vdir$REPLINFO)
 
 ssh -X -a $SEITEB "(
-vserver $SLAVE exec /etc/init.d/mysql stop
-cp -p /etc/vservers/$SLAVE/vdir/var/lib/mysql/master.info /etc/vservers/$SLAVE/vdir/root/
-rm -rf /etc/vservers/$SLAVE/vdir/var/lib/mysql/*
-( cd /mnt/tempmysqlsync && tar cf - . ) | mbuffer -m8G | ( cd /etc/vservers/$SLAVE/vdir/var/lib/mysql && tar xvpf - )
-cp -p /etc/vservers/$SLAVE/vdir/root/master.info /etc/vservers/$SLAVE/vdir/var/lib/mysql/
-vserver $SLAVE exec mysqld_safe --skip-slave-start &
-while ! echo 'SELECT NOW()' | vserver $SLAVE exec mysql; do sleep 1; done
-echo \"SHOW SLAVE STATUS; RESET SLAVE; SHOW SLAVE STATUS; CHANGE MASTER TO MASTER_LOG_FILE='$MASTERLOGFILE', MASTER_LOG_POS=$MASTERLOGPOS; SHOW SLAVE STATUS;\" | vserver $SLAVE exec mysql
-vserver $SLAVE exec /etc/init.d/mysql stop
+	vserver $SLAVE exec /etc/init.d/mysql stop
+	cp -p /etc/vservers/$SLAVE/vdir/var/lib/mysql/master.info /etc/vservers/$SLAVE/vdir/root/
+	rm -rf /etc/vservers/$SLAVE/vdir/var/lib/mysql/*
+)"
+if [ x$NOTMPLV == x0 ]; then
+	ssh -X -a $SEITEB "( ( cd /mnt/tempmysqlsync && tar cf - . ) | mbuffer -m8G | ( cd /etc/vservers/$SLAVE/vdir/var/lib/mysql && tar xvpf - )"
+else
+  rsync -rlHpogDtvxXA --numeric-ids /mnt/tempmysqlsnap/$RELPATH/. $SEITEB:/etc/vservers/$SLAVE/vdir/var/lib/mysql/
+fi
+ssh -X -a $SEITEB "(
+	cp -p /etc/vservers/$SLAVE/vdir/root/master.info /etc/vservers/$SLAVE/vdir/var/lib/mysql/
+	vserver $SLAVE exec mysqld_safe --skip-slave-start &
+	while ! echo 'SELECT NOW()' | vserver $SLAVE exec mysql; do sleep 1; done
+	echo \"SHOW SLAVE STATUS; RESET SLAVE; SHOW SLAVE STATUS; CHANGE MASTER TO MASTER_LOG_FILE='$MASTERLOGFILE', MASTER_LOG_POS=$MASTERLOGPOS; SHOW SLAVE STATUS;\" | vserver $SLAVE exec mysql
+	vserver $SLAVE exec /etc/init.d/mysql stop
 )"
 ssh -X -a $SEITEB "(vserver $SLAVE exec /etc/init.d/mysql restart)"
-ssh -X -a $SEITEB "(
-umount /mnt/tempmysqlsync
-rmdir /mnt/tempmysqlsync
-rmdir /mnt/tempmysqlsnap
-lvremove -f /dev/$TMPVG/$TMPLV
-)"
+if [ x$DRBD == x1 ]; then
+	ssh -X -a $SEITEB "(
+		if [ x$NOTMPLV == x1 ]; then
+		  umount /mnt/tempmysqlsnap &&
+		  lvremove -f /dev/$VGSEITEB/tempmysqlsnap
+		else
+			umount /mnt/tempmysqlsync
+			rmdir /mnt/tempmysqlsync
+			lvremove -f /dev/$TMPVG/$TMPLV
+		fi
+		rmdir /mnt/tempmysqlsnap
+	)"
+else
+	if [ x$NOTMPLV == x1 ]; then
+		umount /mnt/tempmysqlsnap &&
+		lvremove -f /dev/$VGSEITEA/tempmysqlsnap
+	else
+		umount /mnt/tempmysqlsync
+		rmdir /mnt/tempmysqlsync
+		lvremove -f /dev/$TMPVG/$TMPLV
+	fi
+	rmdir /mnt/tempmysqlsnap
+fi
 # vim: ts=2 sw=2 sts=2 sr noet
